## Release Note

추후 입력

+++ src/main/java/com/umc/product/notice/application/port/in/query/dto/NoticeInfo.java
@@ -21,7 +21,7 @@ public record NoticeInfo(
     NoticeTargetInfo targetInfo,
 
     // 메타데이터
-    Integer viewCount,
+    long viewCount,
     Instant createdAt
 ) {
